FUSETOOLS-3669 - Move from non-typed *Properties to typed *Properties for databinding
non-typed *Properties are deprecated for severla years and is planned
for removal
