+++ editor/plugins/org.fusesource.ide.camel.editor/src/org/fusesource/ide/camel/editor/properties/AdvancedEndpointPropertiesSection.java
@@ -22,7 +22,7 @@ import org.eclipse.core.databinding.validation.ValidationStatus;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.jface.databinding.fieldassist.ControlDecorationSupport;
 import org.eclipse.jface.databinding.swt.ISWTObservableValue;
-import org.eclipse.jface.databinding.swt.WidgetProperties;
+import org.eclipse.jface.databinding.swt.typed.WidgetProperties;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.custom.CCombo;
 import org.eclipse.swt.custom.CTabFolder;
@@ -185,7 +185,7 @@ public class AdvancedEndpointPropertiesSection extends FusePropertySection {
 				//initialize the map entry
 				modelMap.put(p.getName(), choiceCombo.getText());
 				// create observables for the control
-				uiObservable = WidgetProperties.selection().observe(choiceCombo);                
+				uiObservable = WidgetProperties.ccomboSelection().observe(choiceCombo);                
 				if (PropertiesUtils.isRequired(p)) {
 					validator = new IValidator() {
 						@Override
